organize csv functionality into a sub module and add facade pattern for all imports
