+++ src/datagrunt/core/ai/engines.py
@@ -9,7 +9,7 @@ from google import genai
 from google.genai import types
 
 @dataclass
-class EngineProperties:
+class AIEngineProperties:
     """Base properties for AI Engines."""
     valid_engines: tuple = ('google')
