+++ tests/core_tests/test_csvcomponents.py
@@ -1,10 +1,10 @@
 from datagrunt.core import (
     CSVDelimiter,
     CSVDialect,
-    CSVRows,
     CSVColumns,
     CSVColumnNameNormalizer,
     CSVComponents,
+    CSVRows,
     CSVStringSample
 )
